% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area Under the Curve}
\usage{
auc(x, ...)
}
\description{
Find the area under the Receiver-Operator and the Precision-Recall curve
}
\examples{

df <-
  data.frame(
      truth = c(1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0)
    , pred  = c(1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0)
  )

cm1 <- confusion_matrix(df$truth, df$pred, thresholds = 1)
auc(cm1)

cm2 <- confusion_matrix(df$truth, df$pred)
auc(cm2)

# for a glm
mod <- glm(
  formula = spam ~ word_freq_our + word_freq_over + capital_run_length_total
, data = spambase
, family = binomial()
)
auc(mod)

}
