% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin.R
\name{check_comments}
\alias{check_comments}
\title{Check Comments}
\usage{
check_comments(c1, c2)
}
\arguments{
\item{c1}{index (line numbers) for the start delimiter of comments}

\item{c2}{index (line numbers) for the end delimiter of comments}
}
\description{
A more robust check for open/close matching sets of comments in a spin file.
}
