% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_target.R
\name{check_target}
\alias{check_target}
\alias{check_target_mtime}
\title{GNU Make like check if traget(s) need to be updated}
\usage{
check_target_mtime(target, prerequisite, verbose)
}
\arguments{
\item{target}{a character vector of paths to targets}

\item{prerequisite}{a character vector of dependencies}
}
\description{
Use modification time and/or checksums of prerequisites to determine if the
target(s) need to be updated.  Similar logic to that used by GNU Make.
}
