% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion Matrices (Contingency Tables)}
\usage{
confusion_matrix(x, ...)
}
\arguments{
\item{x}{prediction vector}

\item{...}{not currently used}

\item{y}{True Condition vector}

\item{formula}{column (known) ~ row (test) for building the confusion matrix}

\item{data}{environment containing the variables listed in the formula}

\item{boot}{boolean, should bootstrapped confidence intervals for the
sensitivity and specificity be computed?  Defaults to FALSE.}

\item{boot_samples}{number of bootstrapping sample to generate, defaults to
1000L.  Ignored if \code{boot == FALSE}.}

\item{alpha}{100(1-alpha)% confidence intervals for specificity and
sensitivity.  Ignored if \code{boot == FALSE}.}
}
\value{
The sensitivity and specificity functions return numeric values.
\code{confusion_matrix} returns a list with elements:
\itemize{
  \item tab the confusion matrix,
  \item specificity point estimate for specificity,
  \item sensitivity point estimate for sensitivity, 
  \item specificity_ci bootstrapped confidence interval for specificity, and
  \item sensitivity_ci bootstrapped confidence interval for sensitivity.
}
}
\description{
Construction of confusion matrices, accuracy, sensitivity,
specificity, confidence intervals (Wilson's method and (optional
bootstrapping)).
}
\details{
Sensitivity and Specificity:
For the sensitivity and specificity function we expect the 2-by-2 confusion
matrix (contingency table) to be of the form:

\tabular{lccc}{
                    \tab      \tab True \tab Condition \cr
                    \tab      \tab -    \tab +         \cr
Predicted Condition \tab -    \tab TN   \tab FN        \cr
Predicted Condition \tab +    \tab FP   \tab TP        \cr
}
where
\itemize{
  \item TN: True Negative,
  \item FP: False Positive, 
  \item FN: False Negative, and
  \item TP: True Positive.
}

sensitivity = TP / (TP + FN)
specificity = TN / (TN + FP) 

This table set up is the result of using 0/1 or boolean variables. See
examples.
}
\examples{
## Example taken from caret::confusionMatrix
\donttest{
lvs <- c("normal", "abnormal")
truth <- factor(rep(lvs, times = c(86, 258)),
                levels = rev(lvs))
pred <- factor(c(rep(lvs, times = c(54, 32)),
                 rep(lvs, times = c(27, 231))),               
               levels = rev(lvs))

confusion_matrix(pred, truth)
confusion_matrix(pred, truth)$stats

temp <- confusion_matrix(pred, truth, boot = TRUE)
temp$stats
temp$boot_stats
}


}

