% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{qroc_build_data_frame}
\alias{qroc_build_data_frame.default}
\alias{qroc_build_data_frame.glm}
\alias{qprc_build_data_frame}
\title{Deprecated Functions}
\usage{
qroc_build_data_frame(fit, truth = NULL, n_threshold = 200, ...)

\method{qroc_build_data_frame}{default}(fit, truth = NULL, n_threshold = 200, ...)

\method{qroc_build_data_frame}{glm}(fit, truth = NULL, n_threshold = 200, ...)

qprc_build_data_frame(fit, n_threshold = 200, ...)
}
\arguments{
\item{fit}{a \code{glm} fit with \code{family = binomial()}, or predicted
values}

\item{truth}{ignored if \code{fit} is a \code{glm} object.  A vector of
observations, 0/1 or FALSE/TRUE values, of equal length to \code{fit}}

\item{n_threshold}{number of thresholds to use to estimate auroc or auprc}

\item{...}{passed to \code{\link[stats]{predict}}}
}
\description{
Archive of deprecated functions. Some of these might be removed from the
package in later releases.

Deprecated methods for building the data sets needed for plotting roc and prc
plots.  use \code{\link{confusion_matrix}} instead.
}
