% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_fstat}
\alias{extract_fpvalue}
\alias{extract_fpvalue.lm}
\alias{extract_fstat}
\title{Extract Summary stats from regression objects}
\usage{
extract_fstat(x)

extract_fpvalue(x)

\method{extract_fpvalue}{lm}(x)
}
\arguments{
\item{x}{a \code{lm} object}
}
\value{
a character vector of the formatted numbers

formatted p-value from the F-test
}
\description{
A collection of functions for extracting summary statistics and
reporting regression results from \code{lm}, \code{glm} and other regression
objects.
}
\details{
TO DO
}
\examples{
# TODO

}
\seealso{
\code{\link{lm}}
}

