% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regression-summary.R
\name{extract_fstat}
\alias{extract_fstat}
\title{Extract Summary stats from regression objects}
\usage{
extract_fstat(x)
}
\arguments{
\item{x}{a \code{lm} object}
}
\value{
a character vector of the formatted numbers
}
\description{
A collectin of functions for extracting summary statistics and
reporting regression results from \code{lm}, \code{glm} and other regression
objects.
}
\details{
TO DO
}
\examples{
# TODO
}
\seealso{
\code{\link{lm}}
}

