% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean_sd.R
\name{geometric_mean_var_sd}
\alias{geometric_mean_var_sd}
\alias{gmean}
\alias{gvar}
\alias{gsd}
\title{Geometric Mean, Variance, and Standard Deviation}
\usage{
gmean(x, na_rm = FALSE)

gvar(x, na_rm = FALSE)

gsd(x, na_rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na_rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}
}
\value{
a numeric value
}
\description{
Return the geometric mean, variance, and standard deviation,
}
