% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{logit}
\alias{logit}
\alias{invlogit}
\title{Deprecated logit and invlogit}
\usage{
logit(x)

invlogit(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Deprecated logit and invlogit functions.  When originally written they were
faster than the base R \code{\link{qlogis}} and \code{\link{plogis}}.
However, that is no longer the case.
}
