% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pefr}
\alias{pefr}
\title{pefr}
\format{
a data frame with four columns
\tabular{rll}{
[, 1] \tab subject     \tab id number \cr
[, 2] \tab measurement \tab first or second \cr
[, 3] \tab meter       \tab \dQuote{Wright peak flow meter} or \dQuote{Mini Write peak flow meter} \cr
[, 4] \tab pefr        \tab peak expiratory flow rate (liters / min) \cr
}
}
\usage{
pefr
}
\description{
Peak expiratory flow rate data
}
\details{
Peak expiratory flow rate (pefr) data is used for examples within the qwraps2
package.  The data has been transcribed from Bland (1986).

\dQuote{The sample comprised colleagues and family of J.M.B. chosen to give a
wide range of PEFR but in no way representative of any defined population.
Two measurements were made with a Wright peak flow meter and two with a mini
Wright meter, in random order. All measurements were taken by J.M.B., using
the same two instruments. (These data were collected to demonstrate the
statistical method and provide no evidence on the comparability of these two
instruments.) We did not repeat suspect readings and took a single reading as
our measurement of PEFR. Only the first measurement by each method is used to
illustrate the comparison of methods, the second measurements being used in
the study of repeatability.}
}
\references{
Bland, J. Martin, and Douglas G Altman. "Statistical methods for assessing
agreement between two methods of clinical measurement." The lancet 327, no. 8476
(1986): 307-310.
}
\seealso{
\code{vignette('qwraps2-data-sets', package = 'qwraps2')} for
details on the construction of the data set.
}
\keyword{datasets}
