% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qacf.R
\name{qacf}
\alias{qacf}
\title{Autocorrelation plot}
\usage{
qacf(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments pased to stats::acf}
}
\value{
a ggplot.
}
\description{
TO BE CONSTRUCTED
}
\details{
TO DO
}

