% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{qkmplot_bulid_data_frame}
\alias{qkmplot_bulid_data_frame}
\title{Deprecated qkmplot data-frame builder}
\usage{
qkmplot_bulid_data_frame(x)
}
\arguments{
\item{x}{object}
}
\description{
Deprecated misspelled wrapper for \code{\link{qkmplot_build_data_frame}}.
}
