% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qprc.R, R/qroc.R
\name{qprc}
\alias{qprc}
\alias{qprc_build_data_frame}
\title{Precision-Recall Curves}
\usage{
qprc(x, ...)

qprc_build_data_frame(fit, n_threshold = 200, ...)

qprc(x, ...)

qprc_build_data_frame(fit, n_threshold = 200, ...)
}
\arguments{
\item{x}{a \code{glm} fit or \code{data.frame} generated by
\code{qprc_build_data_frame}.}

\item{...}{passed to \code{stats::predict}}

\item{fit}{a \code{glm} fit with \code{family = binomial()}.}

\item{n_threshold}{number of thresholds to test against.}
}
\value{
a ggplot.  Minimal aesthetics have been used so that the user may
modify the graphic as desired with ease.
}
\description{
Construction of Precision-Recall Curves (PRC)
}
\details{
Given a \code{glm} fit with \code{family = "binomial"} (either a log-link or
logit-link should be fine, a data set will be constructed and PRC plots
generated.

The area under the curve (AUC) is determined by a trapezoid approximation.
}
\examples{
# load ggplot2 and the diamonds data set
data(diamonds, package = "ggplot2")

# Create a logistic regression models
fit <- glm(I(price > 2800) ~ cut * color, data = diamonds, family = binomial())

head(qprc_build_data_frame(fit))
auc(qprc_build_data_frame(fit))
qprc(fit)

}
\seealso{
\code{vignette("qwraps2-graphics", package = "qwraps2")} for more
examples.
}
