% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qroc.R
\name{qroc-qprc}
\alias{qroc-qprc}
\alias{qroc}
\alias{qroc.default}
\alias{qroc.qwraps2_auc}
\alias{qroc.qwraps2_confusion_matrix}
\alias{qroc.glm}
\alias{qprc}
\alias{qprc.default}
\alias{qprc.qwraps2_auc}
\alias{qprc.qwraps2_confusion_matrix}
\alias{qprc.glm}
\alias{qroc_build_data_frame}
\alias{qroc_build_data_frame.default}
\alias{qroc_build_data_frame.glm}
\alias{qprc_build_data_frame}
\title{Receiver-Operator and Precision-Recall Curves}
\usage{
qroc(x, ...)

\method{qroc}{default}(x, ...)

\method{qroc}{qwraps2_auc}(x, ...)

\method{qroc}{qwraps2_confusion_matrix}(x, ...)

\method{qroc}{glm}(x, ...)

qprc(x, ...)

\method{qprc}{default}(x, ...)

\method{qprc}{qwraps2_auc}(x, ...)

\method{qprc}{qwraps2_confusion_matrix}(x, ...)

\method{qprc}{glm}(x, ...)

qroc_build_data_frame(fit, truth = NULL, n_threshold = 200, ...)

\method{qroc_build_data_frame}{default}(fit, truth = NULL, n_threshold = 200, ...)

\method{qroc_build_data_frame}{glm}(fit, truth = NULL, n_threshold = 200, ...)

qprc_build_data_frame(fit, n_threshold = 200, ...)
}
\arguments{
\item{x}{a \code{glm} fit, \code{qwraps2_confusion_matrix},
\code{qwraps2_auc}, or \code{data.frame}.  In the case of a generic data
frame, there is an expectation of two columns \dQuote{FNR} and \dQuote{TPR} for
\code{qroc}; \dQuote{Recall} and \dQuote{Precision} for \code{qprc}.}

\item{...}{passed to \code{stats::predict}}

\item{fit}{a \code{glm} fit with \code{family = binomial()}, or predicted
values}

\item{truth}{ignored if \code{fit} is a \code{glm} object.  A vector of
observations, 0/1 or FALSE/TRUE values, of equal length to \code{fit}}

\item{n_threshold}{number of thresholds to use to estimate aucroc or auprc}
}
\value{
a ggplot.  Minimal aesthetics have been used so that the user may
modify the graphic as desired with ease.
}
\description{
Construction of ROC and PRC data and plots.
}
\details{
The area under the curve (AUC) is determined by a trapezoid approximation for
both the AUROC and AUPRC.
}
\examples{
#########################################################
# Example 1

df <-
  data.frame(
      truth = c(1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0)
    , pred  = c(1, 1, 1, 1, 1, 1, 0, 0, 1, 0, 0, 0)
  )

cm <- confusion_matrix(df$truth, df$pred)
auc <- auc(cm)
auc$auroc
auc$auprc

# You can get the same ROC/PRC plot from either the confusion_matrix or the
# auc object
qroc(cm)
qroc(auc)
qprc(cm)
qprc(auc)

#########################################################
# Getting a ROC or PRC plot from a glm object:
# you could call confusion_matrix, auc, or just go directly to qroc/qprc

mod <- glm(
  formula = spam ~ word_freq_our + word_freq_over + capital_run_length_total
, data = spambase
, family = binomial()
)

qroc(mod)

qprc(mod)
qprc(mod) + ggplot2::ylim(0, 1)

#########################################################
# plot more than one ROC
mod2 <- update(mod, formula = . ~ word_freq_our)

auc1 <- auc(mod)
auc2 <- auc(mod2)

auroc_data <- rbind(cbind(auc1$roc_data, model = "Model 1"),
                    cbind(auc2$roc_data, model = "Model 2"))

qroc(auroc_data) + ggplot2::aes(color = model)

auc1$auroc
auc2$auroc

auprc_data <- rbind(cbind(auc1$prc_data, model = "Model 1"),
                    cbind(auc2$prc_data, model = "Model 2"))

qprc(auprc_data, prevalence = auc1$prevalence) + ggplot2::aes(color = model)

auc1$auprc
auc2$auprc


}
\seealso{
\code{vignette("qwraps2-graphics", package = "qwraps2")} for more
examples.
}
