% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwraps2-package.R
\docType{package}
\name{qwraps2-package}
\alias{qwraps2}
\alias{qwraps2-package}
\title{Quick Wraps for Reproducible Reports}
\description{
\pkg{qwraps2} is a collection of helper functions for authoring reproducible
analysis reports. The package focuses on formatted summary statistics,
summary tables, regression-result formatting, and \pkg{ggplot2}-based plots.
}
\details{
Several wrappers for \pkg{ggplot2}-style graphics are provided, including
receiver operating characteristic curves, precision-recall curves,
Bland-Altman plots, autocorrelation plots, and Kaplan-Meier plots. These
functions include \code{\link{qroc}}, \code{\link{qprc}},
\code{\link{qacf}}, \code{\link{qblandaltman}}, and \code{\link{qkmplot}}.

Other functions quickly generate meaningful character strings for reporting
results in .Rnw, .Rmd, or similar report files.
}
\section{Vignettes}{

\itemize{
  \item \code{vignette("qwraps2-formatted-summary-statistics", package = "qwraps2")}
  introduces individual summary-statistic and formatting helpers.
  \item \code{vignette("qwraps2-summary-table", package = "qwraps2")}
  demonstrates full summary-table workflows.
  \item \code{vignette("qwraps2-graphics", package = "qwraps2")}
  demonstrates the plotting helpers.
  \item \code{vignette("qwraps2-data-sets", package = "qwraps2")}
  documents the bundled example data sets.
}
}

\section{Options}{

There are several options which can be set via \code{options} and will be
used via \code{getOption}.  The following lists, in alphabetical order the
different options which are available and what they control.

\itemize{
  \item \code{getOption("qwraps2_alpha", 0.05)} significance level, used for
  generating \code{(1 - getOption("qwraps2_alpha", 0.05)) * 100}\% confidence
  intervals, and determining significance for p-value <
  \code{getOption("qwraps2_alpha", 0.05)}.

  \item \code{getOption("qwraps2_frmt_digits", 2)}  Number of digits to the
  right of the decimal point for any value other than p-values.

  \item \code{getOption("qwraps2_frmtp_case", "upper")} set to either
  'upper' or 'lower' for the case of the 'P' for reporting p-values.

  \item \code{getOption("qwraps2_frmtp_digits", 4)}  Number of digits to the
    right of the decimal point to report p-values too.  If
    \code{log10(p-value) < getOption("qwraps2_frmtp_digits", 4)} then the
    output will be "P < 0.01", to however many digits are correct.  Other
    options control other parts of the output p-value format.

  \item \code{getOption("qwraps2_frmtp_leading0", TRUE)} to display or not
  to display the leading zero in p-values, i.e., if TRUE p-values are
  reported as 0.02 versus when FALSE p-values are reported as .02.

  \item \code{getOption("qwraps2_journal", "default")} if a journal has
  specific formatting for p-values or other statistics, this option will
  control the output.  Many other options are ignored if this is any other
  than default.  Check the github wiki, or this file,  for current lists of
  implemented journal style methods.

  \item \code{getOption("qwraps2_markup", "latex")} value set to 'latex' or
  to 'markdown'.  Output is formatted to meet requirements of either markup
  language.
}
}

\section{Journals with predefined formatting}{


\itemize{
  \item Obstetrics & Gynecology
  \itemize{
     \item \url{https://www.editorialmanager.com/ong/default.aspx}
     \item \code{options(qwraps2_journal = "obstetrics_gynecology")}
     \item P-value formatting as of April 2015:

      Express P values to no more than three decimal places.

      Based on observations of published work, leading 0 will be omitted.
   }

  \item Pediatric Dentistry:
  \itemize{
    \item \url{https://www.aapd.org/publications/}
    \item \code{options(qwraps2_journal = "pediatric_dentistry")}
    \item P-value formatting  as of March 2015.

    If P > .01, the actual value for
    P should be expressed to 2 digits.  Non-significant values should not be
    expressed as "NS" whether or not P is significant, unless rounding a
    significant P-value expressed to 3 digits would make it non-significant
    (i.e., P=.049, not P=.05).  If P<.01, it should be expressed to 3 digits
    (e.g., P=.003, not P<.05).  Actual P-values should be expressed unless
    P<.001, in which case they should be so designated.
  }

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dewittpe/qwraps2/}
  \item \url{https://www.peteredewitt.com/qwraps2/}
  \item Report bugs at \url{https://github.com/dewittpe/qwraps2/issues}
}

}
\author{
\strong{Maintainer}: Peter DeWitt \email{dewittpe@gmail.com} (\href{https://orcid.org/0000-0002-6391-0795}{ORCID})

Authors:
\itemize{
  \item Peter DeWitt \email{dewittpe@gmail.com} (\href{https://orcid.org/0000-0002-6391-0795}{ORCID})
}

Other contributors:
\itemize{
  \item Tell Bennett \email{tell.bennett@cuanschutz.edu} (\href{https://orcid.org/0000-0003-1483-4236}{ORCID}) [contributor]
}

}
\keyword{internal}
