% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%s\%}
\alias{\%s\%}
\title{Operators}
\usage{
e1 \%s\% e2
}
\arguments{
\item{e1}{a character string}

\item{e2}{a character string}
}
\description{
A set of helpful operators to make writing and basic data analysis easier.
}
\examples{

# base R
paste0("A longer string ", "can be ", "built")

# with the \%s\% operator
"A longer string " \%s\% "can be " \%s\% "built"

}
