% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qwraps2.R
\docType{package}
\name{qwraps2}
\alias{qwraps2}
\alias{qwraps2-package}
\title{A collection of wrapper functions for aiding the authoring of reproducible
reports.}
\description{
\pkg{qwraps2} is a collection of functions I have found very useful when
working on a varied collection of different analysis reports.  These
functions are commonly used wrappers which I have been reusing and modifying
over several years.
}
\details{
Several wrappers for \pkg{ggplot2} style graphics, such as ROC, AUC, and KM
plots are provided.  Named as \code{\link{qroc}}, \code{\link{qauc}}, and
\code{\link{qkmplot}} to pay homage to \code{qplot} form \pkg{ggplot2} and
the standard names for such plots.

Other functions are used to quickly generate meaningful character strings for
outputting results in .Rnw, .Rmd, or other similar functions.
}

