% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmt.R
\name{Rpkg}
\alias{Rpkg}
\alias{CRANpkg}
\alias{Githubpkg}
\alias{Gitlabpkg}
\title{Formatting Style on URLs for packages on CRAN, Github, and Gitlab.}
\usage{
Rpkg(pkg)

CRANpkg(pkg)

Githubpkg(pkg, username)

Gitlabpkg(pkg, username)
}
\arguments{
\item{pkg}{The name of the package, will work as a quoted or raw name.}

\item{username}{username for Github.com or Gitlab.com}
}
\description{
Functions for controling the look of package names in markdown created
vignettes and easy currating of URLs for the packages.
}
\examples{

Rpkg(qwraps2)
Rpkg("qwraps2")

CRANpkg(qwraps2)
CRANpkg("qwraps2")

Githubpkg(qwraps2, "dewittpe")
Githubpkg("qwraps2", dewittpe)

Gitlabpkg(qwraps2, "dewittpe")
Gitlabpkg("qwraps2", dewittpe)

}
