% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spambase}
\alias{spambase}
\title{Spambase}
\format{
a data.frame with 4601 rows, 58 columns; 57 features and 0/1
indicator for spam

Used under CC BY 4.0 license.
}
\usage{
spambase
}
\description{
Classifying Email as Spam or Non-Spam
}
\references{
Hopkins,Mark, Reeber,Erik, Forman,George, and Suermondt,Jaap. (1999). Spambase. UCI Machine Learning Repository. https://doi.org/10.24432/C53G6X.
}
\seealso{
\code{vignette("qwraps2-data-sets", package = "qwraps2")} for
details on the construction of the data set.
}
\keyword{datasets}
