% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-stepribbon.R
\docType{data}
\name{StatStepribbon}
\alias{StatStepribbon}
\title{Stat Step Ribbon}
\description{
Provides stairstep values for ribbon plots (Copied this from the
https://github.com/hrbrmstr/ggalt version 0.6.0, which is not yet on CRAN.
Some minor modificiations to the file have been made).
}
\references{
\url{https://groups.google.com/forum/?fromgroups=#!topic/ggplot2/9cFWHaH1CPs}
}
\keyword{datasets}
