% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\alias{qsummary}
\alias{cbind.qwraps2_summary_table}
\alias{rbind.qwraps2_summary_table}
\title{Data Summary Tables}
\usage{
summary_table(x, summaries = qsummary(x))

qsummary(x, ...)

\method{cbind}{qwraps2_summary_table}(..., deparse.level = 1)

\method{rbind}{qwraps2_summary_table}(..., deparse.level = 1)
}
\arguments{
\item{x}{a \code{data.frame} or \code{grouped_df}.}

\item{summaries}{a list of lists of formulea for summarizing the data set.
See Details and examples.}

\item{...}{\code{qwraps2_summary_table} objects to bind together}

\item{deparse.level}{integer controlling the construction of labels in the
case of non-matrix-like arguments (for the default method): \code{deparse.level =
0} constructs no labels; the default, \code{deparse.level = 1} or
\code{deparse.level = 2} constructs labels from the argument names.}
}
\value{
a \code{qwraps2_summary_table} object.
}
\description{
Tools useful for building data summary tables.
}
\details{
\code{summary_table} can be used to generate good looking, simple tabels in
LaTeX or markdown.  Functions like xtables::print.xtable and Hmisc::latex
provide many more tools for formating tables.  The purpose of
\code{summary_table} is to generate good looking tables quickly within
workflow for summarizing a data set.

Creating a list-of-lists of summary functions to apply to a data set will
allow the exploration of the whole data set and grouped data sets.  In the
example provided on this page we see a set of summary measures for the
\code{\link[datasets]{mtcars}} data set and the construction of a table for
the whole data set and for a grouped data set.


The list-of-lists should be thought of as follows:  the outer list defines
row groups, the inner lists define the rows within each row group.

More detailed use of these functions can be found the "summary-statistics"
vignette.

The \code{print} method for the \code{qwraps2_summary_table} objects is just
a simple wrapper for \code{\link{qable}}.
}
\seealso{
\code{\link{qable}} for marking up \code{qwraps2_data_summary}
objects.  \code{\link[dplyr]{group_by}} for \code{\link[dplyr]{grouped_df}}
objects.  The \code{vignette("summary-statistics", package = "qwraps2")} for
detailed use of these functions and cavets.

\code{cbind}

\code{rbind}
}
