% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{tab_summary}
\alias{tab_summary}
\title{Tabular Summaries}
\usage{
tab_summary(x, n_perc_args = list(digits = 0, show_symbol = FALSE),
  envir = parent.frame())
}
\arguments{
\item{x}{a variable to summarize}

\item{n_perc_args}{a list of arguments to pass to \code{n_perc}}

\item{envir}{the environment to attach to the resulting formulea}
}
\description{
Tool to quickly generate the code for summarizing a variable.  To be used
with summary_table.
}
