% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmt.R
\name{backtick}
\alias{backtick}
\title{Backtick}
\usage{
backtick(x)
}
\arguments{
\item{x}{the thing to be deparsed and encapsulated in backticks}
}
\description{
Encapsulate a string in backticks. Very helpful for in line code in
knitr::spin scripts.
}
