% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logit}
\alias{invlogit}
\alias{logit}
\title{logit and inverse logit functions}
\usage{
logit(x)

invlogit(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
transform \code{x} either via the logit, or inverse logit.
}
\details{
Why are these functions are not part of base R?  I don't have the
answer to that question.  So here are functions for easy of use.
}

