% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mtcars2}
\alias{mtcars2}
\title{mtcars2}
\format{
An object of class \code{data.frame} with 32 rows and 19 columns.
}
\usage{
mtcars2
}
\description{
An extended version of \code{\link[datasets]{mtcars}} data set.
}
\seealso{
\code{vignette("qwraps2-data-sets", package = "qwraps2")} for
details on the construction of the data set.
}
\keyword{datasets}
