% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qkmplot.R
\name{qkmplot}
\alias{qkmplot}
\alias{qkmplot_bulid_data_frame}
\alias{qkmplot_bulid_data_frame.survfit}
\alias{qrmst}
\alias{qrmst.survfit}
\alias{qrmst.qwraps2_generated}
\title{Kaplan-Meier Plot}
\usage{
qkmplot(x, conf_int = FALSE, ...)

qkmplot_bulid_data_frame(x)

\method{qkmplot_bulid_data_frame}{survfit}(x)

qrmst(x, tau = Inf)

\method{qrmst}{survfit}(x, tau = Inf)

\method{qrmst}{qwraps2_generated}(x, tau = Inf)
}
\arguments{
\item{x}{object}

\item{conf_int}{logical if TRUE show the CI}

\item{...}{Other arguments passed to survival::plot.survfit}

\item{tau}{upper bound on time for restricted mean survial time estimate}
}
\value{
a ggplot.
}
\description{
A ggplot2 version of a Kaplan-Meier Plot
}
\details{
Functions to build, explicitly or implicitly, data.frames and then creating a
ggplot2 KM plot.
}
\examples{
# create a survfit object
require(survival)
leukemia.surv <- survival::survfit(survival::Surv(time, status) ~ x, data = survival::aml)

qkmplot(leukemia.surv, conf_int = TRUE)

qkmplot_bulid_data_frame(leukemia.surv)

qrmst(leukemia.surv) # NaN for rmst.se in Nonmaintained strata as laste observation is an event
qrmst(leukemia.surv, 44) 

}
\seealso{
\code{vignette("qwraps2-graphics", package = "qwraps2")} for
additional examples.
}
