% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qkmplot.R
\name{qkmplot}
\alias{qkmplot}
\title{Kaplan-Meier Plot}
\usage{
qkmplot(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{Other arguments passed to survival::plot.survfit}
}
\value{
a ggplot.
}
\description{
TO BE CONSTRUCTED
}
\details{
TO DO
}

