% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qkmplot.R
\name{qkmplot}
\alias{qkmplot}
\alias{qkmplot_bulid_data_frame}
\title{Kaplan-Meier Plot}
\usage{
qkmplot(x, conf_int = FALSE, ...)

qkmplot_bulid_data_frame(x)
}
\arguments{
\item{x}{object}

\item{conf_int}{logical if TRUE show the CI}

\item{...}{Other arguments passed to survival::plot.survfit}
}
\value{
a ggplot.
}
\description{
TO BE CONSTRUCTED
}
\details{
TO DO
}
\examples{
require(survival)
leukemia.surv <- survival::survfit(survival::Surv(time, status) ~ x, data = survival::aml) 
survival:::plot.survfit(leukemia.surv, conf.int = TRUE, lty = 2:3, col = 1:2)

qkmplot(leukemia.surv, conf_int = TRUE) 




}

