% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwraps2-package.R
\docType{package}
\name{qwraps2-package}
\alias{qwraps2}
\alias{qwraps2-package}
\title{qwraps2: Quick Wraps 2}
\description{
A collection of (wrapper) functions the creator found useful for quickly placing data summaries and formatted regression results into '.Rnw' or '.Rmd' files. Functions for generating commonly used graphics, such as receiver operating curves or Bland-Altman plots, are also provided by 'qwraps2'. 'qwraps2' is a updated version of a package 'qwraps'. The original version 'qwraps' was never submitted to CRAN but can be found at \url{https://github.com/dewittpe/qwraps/}. The implementation and limited scope of the functions within 'qwraps2' \url{https://github.com/dewittpe/qwraps2/} is fundamentally different from 'qwraps'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dewittpe/qwraps2/}
}

}
\author{
\strong{Maintainer}: Peter DeWitt \email{dewittpe@gmail.com} (\href{https://orcid.org/0000-0002-6391-0795}{ORCID})

Other contributors:
\itemize{
  \item Tell Bennett \email{tell.bennett@cuanschutz.edu} (\href{https://orcid.org/0000-0003-1483-4236}{ORCID}) [contributor]
}

}
\keyword{internal}
