% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{cbind.qwraps2_summary_table}
\alias{summary_table}
\alias{tab_summary}
\title{Data Summary Tables}
\usage{
summary_table(.data, summaries)

tab_summary(x)

cbind.qwraps2_summary_table(..., deparse.level = 1)
}
\arguments{
\item{.data}{a \code{data.frame} or \code{grouped_df}.}

\item{summaries}{a list of lists of formulea for summarizing the data set.}

\item{x}{a variable to summarize}

\item{...}{\code{qwraps2_summary_table} objects to bind together}

\item{deparse.level}{integer controlling the construction of labels in the
case of non-matrix-like arguments (for the default method): \code{deparse.level =
0} constructs no labels; the default, \code{deparse.level = 1} or
\code{deparse.level = 2} constructs labels from the argument names.}
}
\value{
a \code{qwraps2_summary_table} object.
}
\description{
Tools useful for building data summary tables.
}
\details{
Detailed use of these functions can be found the a vignette.

The \code{print} method for the \code{qwraps2_summary_table} objects is just
a simple wrapper for \code{qable}.
}
\seealso{
\code{\link{qable}} for marking up \code{qwraps2_data_summary}
objects.  \code{\link[dplyr]{group_by}} for \code{\link[dplyr]{grouped_df}}
objects.

\code{cbind}
}

